regulatory-chatbot-api/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
transformers>=4.35.2
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.5.0
datasets>=4.0.0
accelerate>=0.26.0
